requirements.txt
ipython
odfpy
pyYAML
numpy
pandas
